# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Command-and-control core backend API.",
        "title": "Logos Core API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/wiki/collab-ticket": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Issue a short-lived JWT for Hocuspocus WebSocket authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wiki"
                ],
                "summary": "Get collaboration ticket",
                "parameters": [
                    {
                        "description": "Document to collaborate on",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.collabTicketRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "controller.collabTicketRequest": {
            "type": "object",
            "required": [
                "documentId"
            ],
            "properties": {
                "documentId": {
                    "type": "string"
                },
                "schemaVersion": {
                    "description": "SchemaVersion is the editor schema version compiled into the requesting\nclient. Absent (legacy/old clients that predate this field) decodes to 0.\nCompared against the document's ContentStateSchemaVersion to block clients\ntoo old to safely render — and thus edit — the stored content.",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token (e.g. \"Bearer eyJhb...\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}